Removed artifact uploads and Docker pushes
Discontinued the step for uploading build artifacts for different
platforms in the development workflow. Similarly, removed the Docker
login and push steps in the docker workflow. These changes streamline
the CI/CD process, potentially due to shifting to a different deployment
method or due to redundancy in artifact handling. Future deployments
will require an alternative to the removed steps.

Refs #1432